requirements.txt
django==4.1.6
django-cors-headers==3.13.0
grpcio==1.52.0
grpcio-tools==1.52.0
protobuf==4.21.12